Clean up JUnitHeader and Footer
