--- src/org/labkey/test/tests/JUnitSmokeTest.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2015-2018 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.labkey.test.tests;
-
-import junit.framework.TestSuite;
-import org.junit.experimental.categories.Category;
-import org.labkey.test.categories.Base;
-import org.labkey.test.categories.Smoke;
-
-import java.util.Map;
-
-@Category({Base.class, Smoke.class})
-public class JUnitSmokeTest
-{
-    public static boolean accept(Map<String, Object> test)
-    {
-        return JUnitTest.getCategories(test).contains("smoke");
-    }
-
-    public static TestSuite suite() throws Exception
-    {
-        return JUnitTest._suite(JUnitSmokeTest::accept, true);
-    }
-}
-
