+++ src/org/labkey/test/util/JUnitHeader.java
@@ -19,36 +19,13 @@ import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.labkey.test.BaseWebDriverTest;
-
-import java.util.List;
+import org.labkey.test.tests.JUnitTest;
 
 import static org.labkey.test.WebTestHelper.logToServer;
 
 @BaseWebDriverTest.ClassTimeout(minutes = 3)
-public class JUnitHeader extends BaseWebDriverTest
+public class JUnitHeader extends JUnitTest.BaseJUnitTestWrapper
 {
-    // Used by 'JUnitFooter' to check for leaks from server-side tests
-    static Long startTime = null;
-
-    private static boolean _extraSetup = false;
-
-    public static void enableExtraSetup(boolean runSetup)
-    {
-        JUnitHeader._extraSetup = runSetup;
-    }
-
-    @Override
-    public List<String> getAssociatedModules()
-    {
-        return null;
-    }
-
-    @Override
-    protected String getProjectName()
-    {
-        return null;
-    }
-
     @BeforeClass
     public static void beforeClass()
     {
@@ -58,7 +35,7 @@ public class JUnitHeader extends BaseWebDriverTest
     @Test
     public void configureR()
     {
-        if (_extraSetup)
+        if (extraSetup)
         {
             RReportHelper reportHelper = new RReportHelper(this);
             reportHelper.ensureRConfig(); // reportTest.js (via RhinoService) executes an R script
@@ -68,7 +45,7 @@ public class JUnitHeader extends BaseWebDriverTest
     @Test
     public void configurePipeline()
     {
-        if (_extraSetup)
+        if (extraSetup)
         {
             PipelineToolsHelper pipelineToolsHelper = new PipelineToolsHelper(this);
             pipelineToolsHelper.setToolsDirToTestDefault(); // Point to extra tools if present (currently only sequeneanalysis tools)
@@ -86,7 +63,7 @@ public class JUnitHeader extends BaseWebDriverTest
     @Test
     public void startSystemMaintenance()
     {
-        if (_extraSetup)
+        if (extraSetup)
         {
             super.startSystemMaintenance();
         }
@@ -97,16 +74,5 @@ public class JUnitHeader extends BaseWebDriverTest
     {
         startTime = System.currentTimeMillis();
         logToServer("=== Starting Server-side JUnit Tests ===");
-    }
-
-    @Override
-    protected void checkLinks()
-    {
-        // skip
-    }
-
-    @Override public BrowserType bestBrowser()
-    {
-        return BrowserType.CHROME;
     }
 }
